getting album to display but not pictures yet
