+++ dan-stineback/app/js/first/controllers/index.js
@@ -2,5 +2,5 @@
 
 module.exports = function(app) {
   require('./imageController')(app);
-  require('./photo_controller')(app);
+  require('./photoController')(app);
 };
